services/detector/requirements.txt
kafka-python==2.0.2
redis==4.5.4
scikit-learn==1.2.2
numpy==1.24.3
scipy==1.10.1